Rename Makefile.proto to avoid interpretation as proto file (#1468)
The main reason for the change is I'm trying to use this library with Bazel,
using gazelle. Gazelle automates creation of the Bazel rules for a Go module.
Gazelle assumes files ending in '.proto' are Protocol Buffer files and invokes
protoc on the file to generate the Go code for the proto file. Unfortunately,
protoc doesn't seem to like Makefiles.

As a side-benefit, editors that don't support Emacs mode hints
`-*- mode: makefile; -*-` won't try to highlight the Makefile as a protobuf
file.