Fix flaky OTLP exporter reconnect test (#1814)
* Fix flaky OTLP exporter reconnect test

The tests that check the OTLP exporter will reconnect wait for the
reconnect loop to complete, in theory. However, they do not yield the
active goroutine scheduling. The reconnect loop is in its own goroutine
meaning it is unlikely for that loop to be re-entered, especially on
slow systems. This updates the tests call runtime.Gosched when waiting
for the reconnect loop and yield the scheduling to other goroutines.

* Add changes to changelog

* Use time.After instead of Timer

* Remove changelog entry